booking_service/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
mysqlclient
requests